api/requirements.txt
azure-functions
azure-storage-blob>=12.19.0
pymupdf>=1.24.0
python-docx>=1.1.0
openai>=1.40.0
